RidePlanFramework/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==2.6.0
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
cryptography==37.0.4
decorator==5.1.1
et-xmlfile==1.1.0
execnet==1.9.0
h11==0.13.0
idna==2.10
iniconfig==1.1.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.2
pytest-xdist==2.5.0
requests==2.24.0
selenium==4.3.0
self==2020.12.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
switcher==0.1.8
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.25.11
wsproto==1.1.0
